fix: variable usage with enhanced for loop (#535)
