+++ jadx-core/src/main/java/jadx/core/dex/attributes/AFlag.java
@@ -18,6 +18,7 @@ public enum AFlag {
 	DONT_RENAME, // do not rename during deobfuscation
 	REMOVE, // can be completely removed
 	ADDED_TO_REGION,
+	ITERATOR_IN_FOR, // the iterator in 'for each loops'
 
 	FINALLY_INSNS,
